zeal-rest/src/main/java/com/dzone/albanoj2/zeal/rest/resource/ErrorResource.java
package com.dzone.albanoj2.zeal.rest.resource;

/**
 * Resource representing an error that occurred during a REST API request.
 * 
 * @author Justin Albano <justin.albano.author@gmail.com>
 *
 * @since 1.0.0
 */
public class ErrorResource {

	private String timestamp;
	private String message;
	private String detail;

	public String getTimestamp() {
		return timestamp;
	}

	public void setTimestamp(String timestamp) {
		this.timestamp = timestamp;
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}

	public String getDetail() {
		return detail;
	}

	public void setDetail(String detail) {
		this.detail = detail;
	}
}
